tests/pyproject.toml
[project]
name = "azure-functions-bundle-e2e"
version = "1.0.0"
description = "End To End Testing for Extension Bundle"
authors = [
    { name = "Azure Functions team at Microsoft Corp.", email = "azurefunctions@microsoft.com" }
]
keywords = ["azure", "functions", "azurefunctions",
            "python", "serverless"]
license = { name = "MIT", file = "LICENSE" }
readme = { file = "README.md", content-type = "text/markdown" }
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: MacOS :: MacOS X",
    "Environment :: Web Environment",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers"
]
dependencies = [
    "azure-functions==1.24.0b3",
     "python-dateutil~=2.9.0",
     "protobuf~=4.25.3; python_version >= '3.8' and python_version < '3.13'",
     "protobuf~=5.29.0; python_version >= '3.13'",
     "grpcio-tools~=1.59.0; python_version >= '3.8' and python_version < '3.13'",
     "grpcio-tools~=1.70.0; python_version >= '3.13'",
     "grpcio~=1.43.0; python_version == '3.7'",
     "grpcio ~=1.59.0; python_version >= '3.8' and python_version < '3.13'",
     "grpcio~=1.70.0; python_version >= '3.13'",
     "azurefunctions-extensions-base; python_version >= '3.8'",
     "azure-functions-runtime==1.0.0a3; python_version >= '3.13'",
     "azure-functions-runtime-v1==1.0.0a2; python_version >= '3.13'"
]

[project.urls]
Documentation = "https://github.com/Azure/azure-functions-extension-bundles?tab=readme-ov-file#overview"
Repository = "https://github.com/Azure/azure-functions-extension-bundles"

[project.optional-dependencies]
dev = [
    "azure-eventhub", # Used for EventHub E2E tests
    "azure-functions-durable", # Used for Durable E2E tests
    "azure-monitor-opentelemetry; python_version >= '3.8'", # Used for Azure Monitor unit tests
    "mysql-connector-python", # Used for MySQL E2E tests
    "pytest~=7.4.4",
    "requests==2.*",
    "coverage",
    "pytest-sugar",    
    "pytest-cov",
    "pytest-xdist",
    "pytest-randomly",
    "pytest-instafail",
    "pytest-rerunfailures",
    "pytest-asyncio",
    "python-dotenv",
    "invoke",
]
test-http-v2 = [
    "azurefunctions-extensions-http-fastapi==1.0.0",
    "ujson"
]
test-deferred-bindings = [
    "azurefunctions-extensions-bindings-blob==1.0.0",
    "azurefunctions-extensions-bindings-eventhub==1.0.0b1"
]

[build-system]
requires = ["setuptools>=62", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools.packages.find]
include = [ "emulator_tests", "utils"
    ]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
known_first_party = ["azure_functions_extension_bundles"]
default_section = "THIRDPARTY"
src_paths = ["azure_functions_extension-bundles"]
skip_glob = [
    "build",
    "dist",
    "env",
    "venv",
    "*venv*",
    "*.egg-info",
    ".tox",
    ".nox",
    ".git",
    "__pycache__",
    ".ci",
    ".devcontainer",
    ".github",
    "eng",
    "pack",
    "*protos/*"
]

[tool.setuptools.dynamic]
version = {attr = "azure_functions_extension_bundles.version.VERSION"}
